+++ kotest-assertions/kotest-assertions-core/src/commonMain/kotlin/io/kotest/matchers/collections/CollectionMatchers.kt
@@ -8,52 +8,6 @@ import io.kotest.matchers.MatcherResult
 import io.kotest.matchers.errorCollector
 import io.kotest.matchers.neverNullMatcher
 
-/**
- * Note that if `this` is empty, this assertion will pass.
- * because there are no elements in it that _do not_ fail the test.
- *
- * See a more detailed explanation of this logic concept in ["Vacuous truth"](https://en.wikipedia.org/wiki/Vacuous_truth) article.
- */
-fun <T> existInOrder(vararg ps: (T) -> Boolean): Matcher<Collection<T>?> = existInOrder(ps.asList())
-
-/**
- * Assert that a collections contains a subsequence that matches the given subsequence of predicates, possibly with
- * values in between.
- *
- * Note that if `this` is empty, this assertion will pass.
- * because there are no elements in it that _do not_ fail the test.
- *
- * See a more detailed explanation of this logic concept in ["Vacuous truth"](https://en.wikipedia.org/wiki/Vacuous_truth) article.
- */
-fun <T> existInOrder(predicates: List<(T) -> Boolean>): Matcher<Collection<T>?> = neverNullMatcher { actual ->
-   require(predicates.isNotEmpty()) { "predicates must not be empty" }
-
-   var subsequenceIndex = 0
-   val actualIterator = actual.iterator()
-
-   while (actualIterator.hasNext() && subsequenceIndex < predicates.size) {
-      if (predicates[subsequenceIndex](actualIterator.next())) subsequenceIndex += 1
-   }
-
-   val passed = subsequenceIndex == predicates.size
-
-   val predicateMatchedOutOfOrderDescription = {
-      val predicateMatchedOutOfOrderIndexes = if (passed) emptyList() else {
-         actual.mapIndexedNotNull { index, element ->
-            if (predicates[subsequenceIndex](element)) index else null
-         }
-      }
-      if (predicateMatchedOutOfOrderIndexes.isEmpty()) "" else
-         ",\nbut found element(s) matching the predicate out of order at index(es): ${predicateMatchedOutOfOrderIndexes.print().value}"
-   }
-
-   MatcherResult(
-      passed,
-      { "${actual.print().value} did not match the predicates in order. Predicate at index $subsequenceIndex did not match.${predicateMatchedOutOfOrderDescription()}" },
-      { "${actual.print().value} should not match the predicates in order" }
-   )
-}
-
 fun <T> haveSize(size: Int): Matcher<Collection<T>> = haveSizeMatcher(size)
 
 /**
